+++ nilearn/experimental/surface/__init__.py
@@ -1,17 +1,17 @@
 """The :mod:`nilearn.experimental.surface` module."""
 
+from nilearn.experimental.surface._datasets import (
+    fetch_destrieux,
+    fetch_nki,
+    load_fsaverage,
+    load_fsaverage_data,
+)
 from nilearn.experimental.surface._surface_image import (
     FileMesh,
     InMemoryMesh,
     Mesh,
     PolyMesh,
     SurfaceImage,
-)
-from nilearn.experimental.surface.datasets import (
-    fetch_destrieux,
-    fetch_nki,
-    load_fsaverage,
-    load_fsaverage_data,
 )
 from nilearn.experimental.surface.maskers import (
     SurfaceLabelsMasker,
